+++ NppCSharpPluginPack/Utils/NanInf.cs
@@ -6,7 +6,7 @@
         /// a/b<br></br>
         /// may be necessary to generate infinity or nan at runtime
         /// to avoid the compiler pre-computing things<br></br>
-        /// since if the compiler sees literal 1d/0d in the code
+        /// since if the compiler sees literal division by 0d in the code
         /// it just pre-computes it at compile time
         /// </summary>
         /// <param name="a"></param>